functions.c update: print_number now get long number

+++ functions.c
@@ -51,7 +51,7 @@ int convert_base(int base, long int number, int band)
  *
  * Return: none - void function
  */
-int print_number(int n)
+int print_number(long int n)
 {
 	unsigned int r;
 	int _length = 0;
